Use uint8_t as mouse coordinate type in input parsing.
Since the buffer is char, which is signed in most cases (to be precise the
sign is not specified by the C standard), it leads to bad "overflow" like
behaviour and reports negative mouse coordinates. Using uint8_t solves the
problem. Keep in mind that mouse coordinates are still limited by the single
byte minus 33 positions. Which is [0; 222]. That's a limitation of X10 mouse
input protocol we're using.

Fixes #54.

+++ src/input.inl
@@ -42,8 +42,8 @@ static int parse_escape_seq(struct tb_event *event, const char *buf, int len)
 		event->type = TB_EVENT_MOUSE; // TB_EVENT_KEY by default
 
 		// the coord is 1,1 for upper left
-		event->x = buf[4] - 1 - 32;
-		event->y = buf[5] - 1 - 32;
+		event->x = (uint8_t)buf[4] - 1 - 32;
+		event->y = (uint8_t)buf[5] - 1 - 32;
 
 		return 6;
 	}
